PDJB-670: move scenario logic into factory
