+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/propertyRegistration/EpcRegistrationCyaSummaryRowsFactory.kt
@@ -2,7 +2,12 @@ package uk.gov.communities.prsdb.webapp.journeys.propertyRegistration
 
 import uk.gov.communities.prsdb.webapp.journeys.Destination
 import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.states.EpcState
+import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.EpcAgeCheckMode
+import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.EpcEnergyRatingCheckMode
+import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.EpcInDateAtStartOfTenancyCheckMode
 import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.EpcScenario
+import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.HasEpcMode
+import uk.gov.communities.prsdb.webapp.journeys.shared.Complete
 import uk.gov.communities.prsdb.webapp.models.viewModels.summaryModels.SummaryCardActionViewModel
 import uk.gov.communities.prsdb.webapp.models.viewModels.summaryModels.SummaryListRowViewModel
 import uk.gov.communities.prsdb.webapp.services.EpcCertificateUrlProvider
@@ -10,8 +15,91 @@ import uk.gov.communities.prsdb.webapp.services.EpcCertificateUrlProvider
 class EpcRegistrationCyaSummaryRowsFactory(
     private val epcCertificateUrlProvider: EpcCertificateUrlProvider,
     private val state: EpcState,
-    private val scenario: EpcScenario,
 ) {
+    private val scenario: EpcScenario = determineScenario(state)
+
+    private fun determineScenario(state: EpcState): EpcScenario {
+        val isOccupied = state.isOccupied == true
+        return when {
+            state.hasEpcStep.outcome == HasEpcMode.PROVIDE_LATER -> {
+                if (isOccupied) EpcScenario.SKIPPED_OCCUPIED else EpcScenario.SKIPPED_UNOCCUPIED
+            }
+
+            state.acceptedEpc == null -> {
+                determineNoEpcScenario(state, isOccupied)
+            }
+
+            else -> {
+                determineEpcPresentScenario(state, isOccupied)
+            }
+        }
+    }
+
+    private fun determineNoEpcScenario(
+        state: EpcState,
+        isOccupied: Boolean,
+    ): EpcScenario =
+        when {
+            state.epcExemptionStep.outcome == Complete.COMPLETE -> EpcScenario.NO_EPC_EXEMPT
+            isOccupied -> EpcScenario.NO_EPC_NO_EXEMPTION_OCCUPIED
+            else -> EpcScenario.NO_EPC_NO_EXEMPTION_UNOCCUPIED
+        }
+
+    private fun determineEpcPresentScenario(
+        state: EpcState,
+        isOccupied: Boolean,
+    ): EpcScenario {
+        val isExpired = state.epcAgeCheckStep.outcome == EpcAgeCheckMode.EPC_OLDER_THAN_10_YEARS
+        return if (!isExpired) {
+            determineNotExpiredEpcScenario(state, isOccupied)
+        } else {
+            determineExpiredEpcScenario(state, isOccupied)
+        }
+    }
+
+    private fun determineNotExpiredEpcScenario(
+        state: EpcState,
+        isOccupied: Boolean,
+    ): EpcScenario {
+        if (state.epcEnergyRatingCheckStep.outcome != EpcEnergyRatingCheckMode.EPC_LOW_ENERGY_RATING) {
+            return EpcScenario.VALID_EPC
+        }
+        return when {
+            state.meesExemptionStep.outcome == Complete.COMPLETE -> EpcScenario.LOW_ENERGY_EPC_MEES_EXEMPTION
+            isOccupied -> EpcScenario.LOW_ENERGY_EPC_NO_EXEMPTION_OCCUPIED
+            else -> EpcScenario.LOW_ENERGY_EPC_NO_EXEMPTION_UNOCCUPIED
+        }
+    }
+
+    private fun determineExpiredEpcScenario(
+        state: EpcState,
+        isOccupied: Boolean,
+    ): EpcScenario {
+        if (!isOccupied) return EpcScenario.EPC_EXPIRED_UNOCCUPIED
+        return when (state.epcInDateAtStartOfTenancyCheckStep.outcome) {
+            EpcInDateAtStartOfTenancyCheckMode.NOT_IN_DATE -> {
+                EpcScenario.EPC_EXPIRED_NOT_IN_DATE_OCCUPIED
+            }
+
+            EpcInDateAtStartOfTenancyCheckMode.IN_DATE -> {
+                if (state.epcEnergyRatingCheckStep.outcome != EpcEnergyRatingCheckMode.EPC_LOW_ENERGY_RATING) {
+                    EpcScenario.EPC_EXPIRED_IN_DATE_OCCUPIED
+                } else if (state.meesExemptionStep.outcome == Complete.COMPLETE) {
+                    EpcScenario.LOW_ENERGY_EPC_EXPIRED_IN_DATE_MEES_EXEMPTION_OCCUPIED
+                } else {
+                    EpcScenario.LOW_ENERGY_EPC_EXPIRED_IN_DATE_NO_EXEMPTION_OCCUPIED
+                }
+            }
+
+            null -> {
+                throw IllegalStateException(
+                    "CheckEpcAnswersStep is not reachable for an occupied property " +
+                        "with an expired EPC before the tenancy check is answered",
+                )
+            }
+        }
+    }
+
     fun createEpcCardTitle(): String? = if (isEpcCardShown()) "propertyCompliance.epcTask.checkEpcAnswers.epc.yourEpc" else null
 
     fun createEpcCardActions(): List<SummaryCardActionViewModel>? {
