+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/propertyRegistration/steps/CheckEpcAnswersStepConfig.kt
@@ -16,8 +16,7 @@ class CheckEpcAnswersStepConfig(
     override val formModelClass = NoInputFormModel::class
 
     override fun getStepSpecificContent(state: EpcState): Map<String, Any?> {
-        val scenario = determineScenario(state)
-        val factory = EpcRegistrationCyaSummaryRowsFactory(epcCertificateUrlProvider, state, scenario)
+        val factory = EpcRegistrationCyaSummaryRowsFactory(epcCertificateUrlProvider, state)
         return mapOf(
             "epcCardTitle" to factory.createEpcCardTitle(),
             "epcCardActions" to factory.createEpcCardActions(),
@@ -34,88 +33,6 @@ class CheckEpcAnswersStepConfig(
     override fun chooseTemplate(state: EpcState) = "forms/checkEpcAnswersForm"
 
     override fun mode(state: EpcState) = getFormModelFromStateOrNull(state)?.let { Complete.COMPLETE }
-
-    private fun determineScenario(state: EpcState): EpcScenario {
-        val isOccupied = state.isOccupied == true
-        return when {
-            state.hasEpcStep.outcome == HasEpcMode.PROVIDE_LATER -> {
-                if (isOccupied) EpcScenario.SKIPPED_OCCUPIED else EpcScenario.SKIPPED_UNOCCUPIED
-            }
-
-            state.acceptedEpc == null -> {
-                determineNoEpcScenario(state, isOccupied)
-            }
-
-            else -> {
-                determineEpcPresentScenario(state, isOccupied)
-            }
-        }
-    }
-
-    private fun determineNoEpcScenario(
-        state: EpcState,
-        isOccupied: Boolean,
-    ): EpcScenario =
-        when {
-            state.epcExemptionStep.outcome == Complete.COMPLETE -> EpcScenario.NO_EPC_EXEMPT
-            isOccupied -> EpcScenario.NO_EPC_NO_EXEMPTION_OCCUPIED
-            else -> EpcScenario.NO_EPC_NO_EXEMPTION_UNOCCUPIED
-        }
-
-    private fun determineEpcPresentScenario(
-        state: EpcState,
-        isOccupied: Boolean,
-    ): EpcScenario {
-        val isExpired = state.epcAgeCheckStep.outcome == EpcAgeCheckMode.EPC_OLDER_THAN_10_YEARS
-        return if (!isExpired) {
-            determineNotExpiredEpcScenario(state, isOccupied)
-        } else {
-            determineExpiredEpcScenario(state, isOccupied)
-        }
-    }
-
-    private fun determineNotExpiredEpcScenario(
-        state: EpcState,
-        isOccupied: Boolean,
-    ): EpcScenario {
-        if (state.epcEnergyRatingCheckStep.outcome != EpcEnergyRatingCheckMode.EPC_LOW_ENERGY_RATING) {
-            return EpcScenario.VALID_EPC
-        }
-        return when {
-            state.meesExemptionStep.outcome == Complete.COMPLETE -> EpcScenario.LOW_ENERGY_EPC_MEES_EXEMPTION
-            isOccupied -> EpcScenario.LOW_ENERGY_EPC_NO_EXEMPTION_OCCUPIED
-            else -> EpcScenario.LOW_ENERGY_EPC_NO_EXEMPTION_UNOCCUPIED
-        }
-    }
-
-    private fun determineExpiredEpcScenario(
-        state: EpcState,
-        isOccupied: Boolean,
-    ): EpcScenario {
-        if (!isOccupied) return EpcScenario.EPC_EXPIRED_UNOCCUPIED
-        return when (state.epcInDateAtStartOfTenancyCheckStep.outcome) {
-            EpcInDateAtStartOfTenancyCheckMode.NOT_IN_DATE -> {
-                EpcScenario.EPC_EXPIRED_NOT_IN_DATE_OCCUPIED
-            }
-
-            EpcInDateAtStartOfTenancyCheckMode.IN_DATE -> {
-                if (state.epcEnergyRatingCheckStep.outcome != EpcEnergyRatingCheckMode.EPC_LOW_ENERGY_RATING) {
-                    EpcScenario.EPC_EXPIRED_IN_DATE_OCCUPIED
-                } else if (state.meesExemptionStep.outcome == Complete.COMPLETE) {
-                    EpcScenario.LOW_ENERGY_EPC_EXPIRED_IN_DATE_MEES_EXEMPTION_OCCUPIED
-                } else {
-                    EpcScenario.LOW_ENERGY_EPC_EXPIRED_IN_DATE_NO_EXEMPTION_OCCUPIED
-                }
-            }
-
-            null -> {
-                throw IllegalStateException(
-                    "CheckEpcAnswersStep is not reachable for an occupied property " +
-                        "with an expired EPC before the tenancy check is answered",
-                )
-            }
-        }
-    }
 }
 
 enum class EpcScenario {
